fix!: allow routes to be overridden at config time instead of supporting only hot reloading
BREAKING CHANGE: custom routes are expected in "routes" directory instead of "camel" directory
