+++ src/main/resources/application.properties
@@ -45,5 +45,6 @@ org.unit.id.scheme=ID
 
 camel.springboot.main-run-controller=true
 camel.springboot.routes-reload-enabled=true
-camel.springboot.routes-reload-directory=camel
+camel.springboot.routes-reload-directory=routes
 camel.springboot.routes-reload-remove-all-routes=false
+camel.springboot.routes-include-pattern=file:routes/*
