+++ src/test/resources/application-test.properties
@@ -1,4 +1,2 @@
 dhis2.api.username = admin
 dhis2.api.password = district
-
-camel.springboot.routes-reload-directory=target/test-classes/camel
